PAR ET OU PIPE WORKING + SIGNAL EXEC + SIGNAL HEREDOC

+++ srcs/heredoc.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   heredoc.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: garance <garance@student.42.fr>            +#+  +:+       +#+        */
+/*   By: galambey <galambey@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/13 11:31:49 by galambey          #+#    #+#             */
-/*   Updated: 2023/12/02 08:45:11 by garance          ###   ########.fr       */
+/*   Updated: 2023/12/05 17:16:37 by galambey         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -71,14 +71,26 @@ static int	ft_recover_prompt(int fd_temp, char *lim, t_list *here_doc)
 {
 	char	*line;
 
+	ft_signal_handler_msh_ter();
 	write(1, "> ", 2);
-	line = ft_magic_malloc(ADD, 0, get_next_line(STDIN_FILENO), NO_ENV);
+	line = get_next_line(STDIN_FILENO);
+
+	// write(1, "> ", 2);
+	// line = ft_magic_malloc(ADD, 0, readline(">"), NO_ENV);
+	if (status == 130)
+	{
+		free(line);
+		open("/dev/stdout", O_RDONLY);
+		return (130);
+	}
 	if (!line)
 	{
+		free(line);
 		(close(fd_temp), write(2, "get_next_line: error\n", 21), ft_magic_malloc(FREE, 0, lim, NO_ENV));
 		(unlink(here_doc->content)/*, ft_magic_malloc(FREE, 0, p->name_here_doc, NO_ENV)*/); //Penser a free
 		exit(EXIT_FAILURE);
 	}
+	ft_magic_malloc(ADD, 0, line, NO_ENV);
 	if (ft_strcmp(lim, line) == 0)
 	{
 		ft_magic_malloc(FREE, 0, line, NO_ENV);
@@ -93,7 +105,7 @@ Function called in a loop in the parent function
 For each limiter encountered, create a heredoc and recover the stdin to copy it
 	into the file created
 */
-void	ft_prompt(t_msh *msh, t_split *av, t_list **here_doc)
+int	ft_prompt(t_msh *msh, t_split *av, t_list **here_doc)
 {
 	int		fd_temp;
 	char	*lim;
@@ -113,21 +125,33 @@ void	ft_prompt(t_msh *msh, t_split *av, t_list **here_doc)
 		unlink((*here_doc)->content)/* ,free(p->name_here_doc)*/;
 		exit(EXIT_FAILURE);  //PENSER A FREE 
 	}
+	
 	while (1)
 	{
-		if (ft_recover_prompt(fd_temp, lim, *here_doc) == 1)
+		int err ;
+
+		err = ft_recover_prompt(fd_temp, lim, *here_doc);
+		if (err == 1)
 			break ;
+		if (err == 130)
+		{
+			close(fd_temp);
+			return (130);
+		}
 	}
+	// open("/dev/stdout", O_RDONLY);
+	ft_signal_handler_msh();
 	msh->p.prompt = 1;// (SUPPRIMER PROMPT PAS BESOIN)
 	*here_doc = head;
  	(close(fd_temp), ft_magic_malloc(FREE, 0, lim, NO_ENV));
+	return (0);
 }
 
 /*
 In a loop, create all the here_doc and storage in a linked list the limiter
 	followed by the name of the here_doc created
 */
-void	ft_heredoc(t_msh *msh)
+int	ft_heredoc(t_msh *msh)
 {
 	t_split *head;
 	
@@ -136,10 +160,14 @@ void	ft_heredoc(t_msh *msh)
 	while (msh->av)  // A METTRE DANS FT_EXEC EN PREMIER LIEU
 	{
 		if (msh->av->token == HERE_DOC)
-			ft_prompt(msh, msh->av, &msh->p.here_doc);
+		{
+			if (ft_prompt(msh, msh->av, &msh->p.here_doc) == 130)
+				return (130);
+		}
 		msh->av = msh->av->next;
 	}
 	msh->av = head;
+	return (0);
 }
 
 void	ft_unlink_heredoc(t_list *heredoc)
